requirements.txt
django
uuid
django-allauth
python-dotenv
django-cors-headers
pandas
djangorestframework
djangorestframework-simplejwt
psycopg2-binary>=2.9,<3.0
typing
drf-spectacular
celery==5.2.7
django-celery
django-celery-results
django-celery-beat
flower
redis==4.5.4
boto3
pycurl
gunicorn
